pyproject.toml
[project]
name = "govee-mcp-server"
version = "0.1.0"
description = "MCP server to control Govee LED devices"
authors = []
dependencies = [
    "mcp[cli]",
    "govee-api-laggat>=0.2.2",
    "python-dotenv"
]
requires-python = ">=3.10"

[project.scripts]
govee-cli = "govee_mcp_server.cli:cli_main"

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/govee_mcp_server"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_default_fixture_loop_scope = "function"